Server/requirements.txt
paho-mqtt==1.5.1
numpy
opencv-python==4.9.0.80